+++ src/utils.ts
@@ -1,0 +1,8 @@
+'use strict';
+
+export interface PromiseCallbacks {
+    [key: string]: {
+        resolve: (value?: void | PromiseLike<void>) => void,
+        reject: (reason?: any) => void
+    }
+}